+++ app_php.go
@@ -36,7 +36,7 @@ func (a *App) IsPHPRunning(version string) bool {
 
 func (a *App) DownloadPHP(version string) string {
 	err := php.Download(version, func(percent int, totalMB float64) {
-		runtime.EventsEmit(a.ctx, "php:download-progress", map[string]interface{}{"percent": percent, "totalMB": totalMB})
+		runtime.EventsEmit(a.ctx, "php:download-progress:"+version, map[string]interface{}{"percent": percent, "totalMB": totalMB})
 	})
 	if err != nil {
 		return err.Error()